requirements.txt
# Requirements for YOUlise
streamlit
google-generativeai
google-api-python-client
python-dotenv
